disallow acct system revocation to certain sub. user if he doesn't even have access to it

+++ app/Http/Controllers/Subscription/ManageSubscriptionUsersController.php
@@ -111,7 +111,27 @@ class ManageSubscriptionUsersController extends Controller
 
     public function ajaxGetAccountingSystems(Subscription $subscription)
     {
-        return $subscription->accountingSystems;
+        $subscription_access = SubscriptionUser::where('user_id', auth()->id())
+            ->where('subscription_id', $subscription->id)
+            ->firstOrFail();
+
+        $subQuery = DB::table('accounting_system_users')
+                ->select('accounting_system_id as id', DB::raw('COUNT(accounting_system_id) as hasAccess'))
+                ->where('subscription_user_id', $subscription_access->id)
+                ->groupBy('accounting_system_id');
+
+        $result = AccountingSystem::select(
+                'accounting_systems.*',
+                DB::raw('IFNULL(t.hasAccess, 0) as hasAccess')
+            )   
+            ->where('subscription_id', $subscription->id)
+            ->leftJoinSub($subQuery, 't', function($join){
+                $join->on('t.id', '=', 'accounting_systems.id');
+            })
+            ->where('t.hasAccess', '>', 0)
+            ->get();
+
+        return $result;
     }
 
     public function ajaxAddAccess(SubscriptionUser $subscriptionUser, AddAccountingSystemAccessRequest $request)
@@ -152,7 +172,20 @@ class ManageSubscriptionUsersController extends Controller
 
     public function ajaxUpdateUser(SubscriptionUser $subscriptionUser, Request $request)
     {
-        $current = AccountingSystemUser::where('subscription_user_id', $subscriptionUser->id)->get();
+        // $current = AccountingSystemUser::where('subscription_user_id', $subscriptionUser->id)->get();
+
+        $current = AccountingSystemUser::where('subscription_user_id', $subscriptionUser->id)
+            ->leftJoin('accounting_systems', 'accounting_systems.id', '=', 'accounting_system_users.accounting_system_id')
+            ->leftJoin('subscription_users', 'subscription_users.id', '=', 'accounting_system_users.subscription_user_id')
+            ->leftJoin('subscriptions', 'subscriptions.id', '=', 'subscription_users.subscription_id')
+            ->select(
+                'accounting_system_users.id as id', 
+                'subscription_users.id as subscription_user_id',
+                'accounting_systems.id as accounting_system_id', 
+                'accounting_systems.name as accounting_system_name', 
+                'subscriptions.id as subscription_id',
+            )
+            ->get();
         $added = [];
         $reviewed = [];
 
@@ -211,6 +244,34 @@ class ManageSubscriptionUsersController extends Controller
                 }
             }
             if(!$found) {
+                // check if authenticated user has access to this accounting system. if the authenticated user
+                // doesn't, this user's access to this accounting system will remain as is.
+                $subscription_access = SubscriptionUser::where('user_id', auth()->id())
+                    ->where('subscription_id', $current[$i]->subscription_id)
+                    ->first();
+
+                $subQuery = DB::table('accounting_system_users')
+                        ->select('accounting_system_id as id', DB::raw('COUNT(accounting_system_id) as hasAccess'))
+                        ->where('subscription_user_id', $subscription_access->id)
+                        ->groupBy('accounting_system_id');
+
+                $result = AccountingSystem::select(
+                        'accounting_systems.*',
+                        DB::raw('IFNULL(t.hasAccess, 0) as hasAccess')
+                    )   
+                    ->where('subscription_id', $current[$i]->subscription_id)
+                    ->where('accounting_systems.id', $current[$i]->accounting_system_id)
+                    ->leftJoinSub($subQuery, 't', function($join){
+                        $join->on('t.id', '=', 'accounting_systems.id');
+                    })
+                    ->first();
+
+                if(!$result || $result->hasAccess == 0) {
+                    continue;
+                }              
+
+                // if authenticated user actually has access to this accounting system, then he/she can
+                // revoke this user access to this accounting system.
                 $removed[] = $current[$i];
                 $current[$i]->permissions()->delete();
                 $current[$i]->delete();
